Replace org.json with gson
Sadly, org.json does not seem compatible with jlink for deployment, as
it isn't packaged as a proper module. Replace with gson instead.

Note: json is barely used in the java code, only to pass a few values
between the BLE drivers and the front-end, and between WinBLE's java
half and WinBLE's native c-sharp half.