Fobot list UI formatting fixes.
Make list rows alternate backgrounds alternate shading (this was
previously attempted in CSS, but was broken, need to use nth-of-type
instead of nth-child in CSS selector). Make the shading a little darker.

Make list a little wider, to avoid excessive line wraps.

Make list rows a little less tall, to avoid excessive whitespace when
there are several robots in list.